requirements.txt
appdirs==1.4.0
chartjs==1.2
dj-database-url==0.5.0
Django==2.1.2
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-heroku==0.3.1
django-mathfilters==0.4.0
django-recaptcha==1.0.5
djangorestframework==3.8.2
Faker==0.9.2
gunicorn==19.6.0
packaging==16.8
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyparsing==2.1.10
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
six==1.10.0
sqlparse==0.2.4
text-unidecode==1.2
whitenoise==3.2
